Ordnerstruktur an J4-Konvention angepasst
