+++ media/index.html
@@ -1,0 +1,1 @@
+<html><body bgcolor="#FFFFFF"></body></html>
