+++ mod_etliste.php
@@ -8,7 +8,7 @@
 defined('_JEXEC') or die;
 
 // Include the syndicate functions only once
-require_once dirname(__FILE__) . '/helper.php';
+require_once dirname(__FILE__) . '/Helper/GetListeHelper.php';
 $liste = modETListeHelper::getListe($params);
 
 // Modulklassen-Suffix
